feat: fix Account balance to BigDecimal

+++ src/gr/aueb/cf/cf9/ch18/bankapp/dto/AccountDepositDTO.java
@@ -1,6 +1,8 @@
 package gr.aueb.cf.cf9.ch18.bankapp.dto;
 
-public record AccountDepositDTO(String iban, double amount) {
+import java.math.BigDecimal;
+
+public record AccountDepositDTO(String iban, BigDecimal amount) {
 
 
 }
